requirements.txt
selenium==4.25.0
undetected-chromedriver==3.5.5
fpdf2==2.7.9
gspread==6.1.3
oauth2client==4.1.3
python-dotenv==1.0.1
schedule==1.2.2
openpyxl==3.1.5
streamlit==1.42.0
pillow==11.1.0